+++ src/senaite/core/datamanagers/field/sample_analyses.py
@@ -50,7 +50,7 @@ class SampleAnalysesFieldDataManager(FieldDataManager):
         brains = catalog(query)
         if kw.get("full_objects", False):
             return map(api.get_object, brains)
-        return brains
+        return list(brains)
 
     def set(self, items, prices, specs, hidden, **kw):
         """Set/Assign Analyses to this AR
@@ -113,12 +113,6 @@ class SampleAnalysesFieldDataManager(FieldDataManager):
 
         # Remove analyses
         map(self.remove_analysis, to_remove)
-
-        # Store the uids in instance's attribute for this field
-        # Note we only store the UIDs of the contained analyses!
-        contained = self.context.objectValues("Analysis")
-        contained_uids = [analysis.UID() for analysis in contained]
-        self.field.setRaw(self.context, contained_uids)
 
     def resolve_specs(self, instance, results_ranges):
         """Returns a dictionary where the key is the service_uid and the value