+++ src/senaite/core/tests/test_calculations.py
@@ -358,7 +358,7 @@ class TestCalculations(DataTestCase):
 
             # Create the AR
             client = self.portal.clients['client-1']
-            sampletype = self.portal.bika_setup.bika_sampletypes['sampletype-1']
+            sampletype = self.portal.setup.sampletypes['sampletype-1']
             values = {'Client': client.UID(),
                       'Contact': client.getContacts()[0].UID(),
                       'SamplingDate': '2015-01-01',
@@ -437,7 +437,7 @@ class TestCalculations(DataTestCase):
                     service.setPrecision(case['fixed_precision'])
                 # Create the AR
                 client = self.portal.clients['client-1']
-                sampletype = self.portal.bika_setup.bika_sampletypes['sampletype-1']
+                sampletype = self.portal.setup.sampletypes['sampletype-1']
                 values = {'Client': client.UID(),
                           'Contact': client.getContacts()[0].UID(),
                           'SamplingDate': '2015-01-01',
@@ -513,7 +513,7 @@ class TestCalculations(DataTestCase):
                     service.setUncertainties(case['uncertainties'])
                 # Create the AR
                 client = self.portal.clients['client-1']
-                sampletype = self.portal.bika_setup.bika_sampletypes['sampletype-1']
+                sampletype = self.portal.setup.sampletypes['sampletype-1']
                 values = {'Client': client.UID(),
                           'Contact': client.getContacts()[0].UID(),
                           'SamplingDate': '2015-01-01',