+++ src/senaite/core/tests/test_duplicate-analysis.py
@@ -51,7 +51,7 @@ class TestAddDuplicateAnalysis(DataTestCase):
         # SampleType:   Apple Pulp
         # Contact:      Rita Mohale
         client = self.portal.clients['client-1']
-        sampletype = self.portal.bika_setup.bika_sampletypes['sampletype-1']
+        sampletype = self.portal.setup.sampletypes['sampletype-1']
         values = {'Client': client.UID(),
                   'Contact': client.getContacts()[0].UID(),
                   'SamplingDate': '2015-01-01',